requirements.txt
websocket-client==1.8.0
pandas==2.2.2
numpy==1.26.4
matplotlib==3.9.0
